Laplacian Eigenvalues and Eigenspaces Calculation
Compute eigenvalues of Laplace operator given (range of) N and m (number of nodes and number of connected neighbors), and count the dimensions for the eigenspaces (i.e., which and how many eigenvectors share the same eigenvalue). A table(or a number of tables if given a range of N) collecting the information is printed out.